1、贷款平台管理，完成增和改

+++ api/models/Platform.js
@@ -27,7 +27,7 @@ module.exports = {
     },
     grade: {
       type: 'number',
-      columnType:'float',
+      columnType:'int',
     },
     fastestTime: {
       type: 'string'
